% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster_interval}
\alias{cluster_interval}
\title{Cluster ranges which are implemented as 2 equal-length numeric vectors.}
\usage{
cluster_interval(starts, ends, max_distance = 0L)
}
\arguments{
\item{starts}{A numeric vector that defines the starts of each interval}

\item{ends}{A numeric vector that defines the ends of each interval}

\item{max_distance}{The maximum distance up to which intervals are still considered to be
the same cluster. Default: 0.}
}
\description{
Cluster ranges which are implemented as 2 equal-length numeric vectors.
}
\examples{
starts <- c(50, 100, 120)
ends <- c(75, 130, 150)
j <- cluster_interval(starts, ends)
j == c(0,1,1)
}
